Develop delete user

+++ src/users/users.controller.ts
@@ -7,12 +7,18 @@ import {
   Param,
   Delete,
   UseGuards,
+  Request,
+  HttpException,
+  HttpStatus,
+  Logger,
 } from '@nestjs/common';
+import { ApiTags, ApiBearerAuth, ApiResponse } from '@nestjs/swagger';
 import { JwtAuthGuard } from 'src/auth/jwt-auth.guard';
+import { ValidatedJWTUserData } from 'src/auth/jwt.strategy';
 import { UsersService } from './users.service';
 import { CreateUserDto } from './dto/create-user.dto';
 import { UpdateUserDto } from './dto/update-user.dto';
-import { ApiTags, ApiBearerAuth } from '@nestjs/swagger';
+import { UserRole } from './entities/user.entity';
 
 @ApiTags('users')
 @ApiBearerAuth()
@@ -41,8 +47,31 @@ export class UsersController {
     return this.usersService.update(id, updateUserDto);
   }
 
+  // @ApiResponse({
+  //   status: HttpStatus.OK,
+  //   description: 'User deleted.',
+  // })
+  // @ApiResponse({
+  //   status: HttpStatus.FORBIDDEN,
+  //   description: 'Not allowed to delete a user.',
+  // })
+  // @ApiResponse({ status: HttpStatus.NOT_FOUND, description: 'User not found.' })
   @Delete(':id')
-  remove(@Param('id') id: string) {
+  remove(
+    @Param('id') id: string,
+    @Request() req: { user: ValidatedJWTUserData },
+  ) {
+    const { user } = req;
+
+    Logger.log(user);
+
+    if (user.role !== UserRole.ADMIN) {
+      throw new HttpException(
+        'You are not allowed to delete a user.',
+        HttpStatus.FORBIDDEN,
+      );
+    }
+
     return this.usersService.remove(id);
   }
 }
